requirements.txt
pdfminer.six
langchain
langchain-community
langchain-google-genai
PyMuPDF
python-dotenv
faiss-cpu
google-genai
streamlit